requirements.txt
Django==1.10.5
django-enumfields==0.8.2
django-filter==1.0.1
djangorestframework==3.5.4
enum34==1.1.6
Markdown==2.6.8
praw==4.3.0
prawcore==0.7.0
requests==2.13.0
six==1.10.0
update-checker==0.16
